+++ Nemesis-Mod/src/bluetooth/services/BlasterSpeedService.h
