+++ Nemesis-Mod/src/bluetooth/services/ConfigurationService.cpp
@@ -1,0 +1,6 @@
+#include "ConfigurationService.h"
+#include "CustomUuid.h"
+
+ConfigurationService::ConfigurationService() : CustomBLEService(UUID128_SVC_CONFIGURATION) {
+
+}
