+++ Nemesis-Mod/src/bluetooth/services/ConfigurationService.h
@@ -1,0 +1,13 @@
+#ifndef CONFIGURATION_SERVICE_H
+#define CONFIGURATION_SERVICE_H
+
+#include "CustomBLEService.h"
+
+class ConfigurationService : public CustomBLEService {
+    public:
+        ConfigurationService();
+
+        void init() override;
+};
+
+#endif /* CONFIGURATION_SERVICE_H */
