+++ Nemesis-Mod/src/bluetooth/services/CustomBLEService.h
