+++ Nemesis-Mod/src/bluetooth/services/CustomUuid.cpp
@@ -11,8 +11,9 @@
 // *************************************************************************************************************
 
 // Services
-uint8_t UUID128_SVC_NERF_BLASTER[16]       = { 0xf0,0xf1,0x29,0x87,0xd0,0xc4,0xbe,0x47,0xb0,0x95,0xc6,0x63,0x09,0xff,0x17,0x68 };        // 6817ff09-63c6-95b0-47be-c4d08729f1f0
+uint8_t UUID128_SVC_NERF_BLASTER[16]       = { 0xf0,0xf1,0x29,0x87,0xd0,0xc4,0xbe,0x47,0xb0,0x95,0x00,0x00,0x09,0xff,0x17,0x68 };        // 6817ff09-0000-95b0-47be-c4d08729f1f0
+uint8_t UUID128_SVC_CONFIGURATION[16]       = { 0xf0,0xf1,0x29,0x87,0xd0,0xc4,0xbe,0x47,0xb0,0x95,0x01,0x00,0x09,0xff,0x17,0x68 };       // 6817ff09-0001-95b0-47be-c4d08729f1f0
 
 // Characteristics
-uint8_t UUID128_CHR_FLYWHEEL_SPEED[16]     = { 0xf0,0xf1,0x29,0x87,0xd0,0xc4,0xbe,0x47,0xb0,0x95,0xc6,0x63,0x00,0x01,0x00,0x00 };        // 00000100-63c6-95b0-47be-c4d08729f1f0
-uint8_t UUID128_CHR_BELT_SPEED[16]         = { 0xf0,0xf1,0x29,0x87,0xd0,0xc4,0xbe,0x47,0xb0,0x95,0xc6,0x63,0x01,0x01,0x00,0x00 };        // 00000101-63c6-95b0-47be-c4d08729f1f0
+uint8_t UUID128_CHR_FLYWHEEL_SPEED[16]     = { 0xf0,0xf1,0x29,0x87,0xd0,0xc4,0xbe,0x47,0xb0,0x95,0x00,0x00,0x00,0x01,0x00,0x00 };        // 00000100-0000-95b0-47be-c4d08729f1f0
+uint8_t UUID128_CHR_BELT_SPEED[16]         = { 0xf0,0xf1,0x29,0x87,0xd0,0xc4,0xbe,0x47,0xb0,0x95,0x00,0x00,0x01,0x01,0x00,0x00 };        // 00000101-0000-95b0-47be-c4d08729f1f0
